Crate flux_macros

Modules

diagnostics 🔒

Copy of compiler/rustc_macros/src/diagnostics from rustc 8925ea358 (nightly-2026-08-21).

fold 🔒

primops 🔒

symbols 🔒

This is based on the implementation of the symbols macro in rustc.

Macros

msg

Creates a DiagMessage from an inline Fluent pattern. The pattern is checked to be valid Fluent, but variables are not checked against the arguments set on the diagnostic.

primop_rules

symbols

Functions

debug_as_json 🔒

Derive Macros

DebugAsJson

Diagnostic

Subdiagnostic

TypeFoldable

TypeVisitable