flux_macros::diagnostics
Module diagnostic_builder

Structs

DiagnosticDeriveVariantBuilder 🔒

Tracks persistent information required for a specific variant when building up individual calls to diagnostic methods for generated diagnostic derives - both Diagnostic for fatal/errors/warnings and LintDiagnostic for lints.

Enums

DiagnosticDeriveKind 🔒

What kind of diagnostic is being derived - a fatal/error/warning or a lint?