Struct flux_macros::diagnostics::diagnostic_builder::DiagnosticDeriveVariantBuilder

pub(crate) struct DiagnosticDeriveVariantBuilder<'parent> {
    pub parent: &'parent DiagnosticDeriveBuilder,
    pub formatting_init: TokenStream,
    pub span: Span,
    pub field_map: HashMap<String, TokenStream>,
    pub slug: Option<(Path, Span)>,
    pub code: Option<((), Span)>,
}

Tracks persistent information required for a specific variant when building up individual calls to diagnostic methods for generated diagnostic derives - both Diagnostic for fatal/errors/warnings and LintDiagnostic for lints.

Fields
parent: &'parent DiagnosticDeriveBuilder

The parent builder for the entire type.

formatting_init: TokenStream

Initialization of format strings for code suggestions.

span: Span

Span of the struct or the enum variant.

field_map: HashMap<String, TokenStream>

Store a map of field name to its corresponding field. This is built on construction of the derive builder.

slug: Option<(Path, Span)>

Slug is a mandatory part of the struct attribute as corresponds to the Fluent message that has the actual diagnostic message.

code: Option<((), Span)>

Error codes are a optional part of the struct attribute - this is only set to detect multiple specifications.

Implementations

impl<'a> DiagnosticDeriveVariantBuilder<'a>

pub fn preamble(&mut self, variant: &VariantInfo<'_>) -> TokenStream

Generates calls to code and similar functions based on the attributes on the type or variant.

pub fn body(&mut self, variant: &VariantInfo<'_>) -> TokenStream

Generates calls to span_label and similar functions based on the attributes on fields or calls to set_arg when no attributes are present.

fn parse_subdiag_attribute( &self, attr: &Attribute ) -> Result<Option<(SubdiagnosticKind, Path, bool)>, DiagnosticDeriveError>

Parse a SubdiagnosticKind from an Attribute.

fn generate_structure_code_for_attr( &mut self, attr: &Attribute ) -> Result<TokenStream, DiagnosticDeriveError>

Establishes state in the DiagnosticDeriveBuilder resulting from the struct attributes like #[diag(..)], such as the slug and error code. Generates diagnostic builder calls for setting error code and creating note/help messages.

fn generate_field_code(&mut self, binding_info: &BindingInfo<'_>) -> TokenStream

fn generate_field_attrs_code( &mut self, binding_info: &BindingInfo<'_> ) -> TokenStream

fn generate_inner_field_code( &mut self, attr: &Attribute, info: FieldInfo<'_>, binding: TokenStream ) -> Result<TokenStream, DiagnosticDeriveError>

fn add_spanned_subdiagnostic( &self, field_binding: TokenStream, kind: &Ident, fluent_attr_identifier: Path ) -> TokenStream

Adds a spanned subdiagnostic by generating a diag.span_$kind call with the current slug and fluent_attr_identifier.

fn add_subdiagnostic( &self, kind: &Ident, fluent_attr_identifier: Path ) -> TokenStream

Adds a subdiagnostic by generating a diag.span_$kind call with the current slug and fluent_attr_identifier.

fn span_and_applicability_of_ty( &self, info: FieldInfo<'_> ) -> Result<(TokenStream, Option<(TokenStream, Span)>), DiagnosticDeriveError>