flux_macros::diagnostics

Module error

Macros

throw_invalid_attr 🔒

Emit an error diagnostic for an invalid attribute (optionally performing additional decoration using the FnOnce passed in diag) and return Err(ErrorHandled).

throw_span_err 🔒

Emit a diagnostic on span $span with msg $msg (optionally performing additional decoration using the FnOnce passed in diag) and return Err(ErrorHandled).

Enums

DiagnosticDeriveError 🔒

Functions

_throw_err 🔒

Helper function for use with throw_* macros - constraints $f to an impl FnOnce.

invalid_attr 🔒

Returns an error diagnostic for an invalid attribute.

path_to_string 🔒

Helper function for printing syn::Path - doesn’t handle arguments in paths and these are unlikely to come up much in use of the macro.

span_err 🔒

Returns an error diagnostic on span span with msg msg.