flux_macros::diagnostics::error

Function path_to_string

fn path_to_string(path: &Path) -> String

Helper function for printing syn::Path - doesn’t handle arguments in paths and these are unlikely to come up much in use of the macro.