flux_macros::diagnostics
Module subdiagnostic

Structs

KindsStatistics 🔒

Provides frequently-needed information about the diagnostic kinds being derived for this type.

SubdiagnosticDerive 🔒

The central struct for constructing the add_to_diag method from an annotated struct.

SubdiagnosticDeriveVariantBuilder 🔒

Tracks persistent information required for building up the call to add to the diagnostic for the final generated method. This is a separate struct to SubdiagnosticDerive only to be able to destructure and split self.builder and the self.structure up to avoid a double mut borrow later on.