Struct flux_macros::diagnostics::subdiagnostic::SubdiagnosticDeriveVariantBuilder

struct SubdiagnosticDeriveVariantBuilder<'parent, 'a> {
    parent: &'parent SubdiagnosticDerive,
    variant: &'a VariantInfo<'a>,
    span: Span,
    formatting_init: TokenStream,
    fields: HashMap<String, TokenStream>,
    span_field: Option<(Ident, Span)>,
    applicability: Option<(TokenStream, Span)>,
    has_suggestion_parts: bool,
    is_enum: bool,
}

Tracks persistent information required for building up the call to add to the diagnostic for the final generated method. This is a separate struct to SubdiagnosticDerive only to be able to destructure and split self.builder and the self.structure up to avoid a double mut borrow later on.

Fields
parent: &'parent SubdiagnosticDerive

The identifier to use for the generated Diag instance.

variant: &'a VariantInfo<'a>

Info for the current variant (or the type if not an enum).

span: Span

Span for the entire type.

formatting_init: TokenStream

Initialization of format strings for code suggestions.

fields: HashMap<String, TokenStream>

Store a map of field name to its corresponding field. This is built on construction of the derive builder.

span_field: Option<(Ident, Span)>

Identifier for the binding to the #[primary_span] field.

applicability: Option<(TokenStream, Span)>

The binding to the #[applicability] field, if present.

has_suggestion_parts: bool

Set to true when a #[suggestion_part] field is encountered, used to generate an error during finalization if still false.

is_enum: bool

Set to true when this variant is an enum variant rather than just the body of a struct.

Implementations

impl<'parent, 'a> SubdiagnosticDeriveVariantBuilder<'parent, 'a>

fn identify_kind( &mut self ) -> Result<Vec<(SubdiagnosticKind, Path, bool)>, DiagnosticDeriveError>

fn generate_field_arg(&mut self, binding_info: &BindingInfo<'_>) -> TokenStream

Generates the code for a field with no attributes.

fn generate_field_attr_code( &mut self, binding: &BindingInfo<'_>, kind_stats: KindsStatistics ) -> TokenStream

Generates the necessary code for all attributes on a field.

fn generate_field_code_inner( &mut self, kind_stats: KindsStatistics, attr: &Attribute, info: FieldInfo<'_>, clone_suggestion_code: bool ) -> Result<TokenStream, DiagnosticDeriveError>

fn generate_field_code_inner_path( &mut self, kind_stats: KindsStatistics, attr: &Attribute, info: FieldInfo<'_>, path: Path ) -> Result<TokenStream, DiagnosticDeriveError>

Generates the code for a [Meta::Path]-like attribute on a field (e.g. #[primary_span]).

fn generate_field_code_inner_list( &mut self, kind_stats: KindsStatistics, attr: &Attribute, info: FieldInfo<'_>, list: &MetaList, clone_suggestion_code: bool ) -> Result<TokenStream, DiagnosticDeriveError>

Generates the code for a [Meta::List]-like attribute on a field (e.g. #[suggestion_part(code = "...")]).

pub(crate) fn into_tokens( &mut self ) -> Result<TokenStream, DiagnosticDeriveError>