Enum flux_macros::diagnostics::utils::SubdiagnosticKind

pub(super) enum SubdiagnosticKind {
    Label,
    Note,
    Help,
    Warn,
    Suggestion {
        suggestion_kind: SuggestionKind,
        applicability: Option<(Applicability, Span)>,
        code_field: Ident,
        code_init: TokenStream,
    },
    MultipartSuggestion {
        suggestion_kind: SuggestionKind,
        applicability: Option<(Applicability, Span)>,
    },
}

Types of subdiagnostics that can be created using attributes

Variants

Label

#[label(...)]

Note

#[note(...)]

Help

#[help(...)]

Warn

#[warning(...)]

Suggestion

Fields

suggestion_kind: SuggestionKind

applicability: Option<(Applicability, Span)>

code_field: Ident

Identifier for variable used for formatted code, e.g. ___code_0. Enables separation of formatting and diagnostic emission so that set_arg calls can happen in-between..

code_init: TokenStream

Initialization logic for code_field’s variable, e.g. let __formatted_code = /* whatever */;

#[suggestion{,_short,_hidden,_verbose}]

MultipartSuggestion

Fields

suggestion_kind: SuggestionKind

applicability: Option<(Applicability, Span)>

#[multipart_suggestion{,_short,_hidden,_verbose}]