flux_macros::diagnostics

Module fluent

Functions

failed 🔒

Tokens to be returned when the macro cannot proceed.

finish 🔒

Final tokens.

fluent_messages 🔒

invocation_relative_path_to_absolute 🔒

Helper function for returning an absolute path for macro-invocation relative file paths.

variable_references 🔒