Module flux_macros::diagnostics::utils

Structs

FieldInfo 🔒

Field information passed to the builder. Deliberately omits attrs to discourage the generate_* methods from walking the attributes themselves.

SubdiagnosticVariant 🔒

Enums

AllowMultipleAlternatives 🔒

Applicability 🔒

Applicability of a suggestion - mirrors rustc_errors::Applicability - and used to represent the user’s selection of applicability if specified in an attribute.

FieldInnerTy 🔒

Inner type of a field and type of wrapper.

SubdiagnosticKind 🔒

Types of subdiagnostics that can be created using attributes

SuggestionKind 🔒

Possible styles for suggestion subdiagnostics.

Constants

CODE_IDENT_COUNT 🔒

Traits

HasFieldMap 🔒

SetOnce 🔒

Small helper trait for abstracting over Option fields that contain a value and a Span for error reporting if they are set more than once.

Functions

build_field_mapping 🔒

Build the mapping of field names to fields. This allows attributes to peek values from other fields.

build_suggestion_code 🔒

Constructs the format!() invocation(s) necessary for a #[suggestion*(code = "foo")] or #[suggestion*(code("foo", "bar"))] attribute field

is_doc_comment 🔒

new_code_ident 🔒

Returns an ident of the form __code_N where N is incremented once with every call.

parse_suggestion_values 🔒

report_error_if_not_applied_to_applicability 🔒

Reports an error if the field’s type is not Applicability.

report_error_if_not_applied_to_span 🔒

Reports an error if the field’s type is not Span.

report_error_if_not_applied_to_ty 🔒

Reports an error if the field’s type does not match path.

report_type_error 🔒

Reports a type error for field with attr.

should_generate_arg 🔒

Returns true if field should generate a arg call rather than any other diagnostic call (like span_label).

type_is_bool 🔒

Checks whether the type ty is bool.

type_is_unit 🔒

Checks whether the type ty is ().

type_matches_path 🔒

Checks whether the type name of ty matches name.

Type Aliases

FieldMap 🔒

SpannedOption 🔒

An Option<T> that keeps track of the span that caused it to be set; used with SetOnce.