flux_macros::diagnostics::utils
Function type_matches_path

pub(crate) fn type_matches_path(ty: &Type, name: &[&str]) -> bool

Checks whether the type name of ty matches name.

Given some struct at a::b::c::Foo, this will return true for c::Foo, b::c::Foo, or a::b::c::Foo. This reasonably allows qualified names to be used in the macro.