flux_macros::diagnostics::utils

Trait SetOnce

pub(crate) trait SetOnce<T> {
    // Required methods
    fn set_once(&mut self, value: T, span: Span);
    fn value(self) -> Option<T>;
    fn value_ref(&self) -> Option<&T>;
}

Small helper trait for abstracting over Option fields that contain a value and a Span for error reporting if they are set more than once.

Required Methods

fn set_once(&mut self, value: T, span: Span)

fn value(self) -> Option<T>

fn value_ref(&self) -> Option<&T>